Refactoring RESTful routes
